IIA-2083 refactored to move ConfirmationText and ConfiramtionMessages outside of the view model
