+++ kview/src/main/java/module-info.java
@@ -143,10 +143,12 @@ module dev.ikm.komet.kview {
     opens dev.ikm.komet.kview.klwindows to javafx.fxml, org.carlfx.cognitive;
     opens dev.ikm.komet.kview.mvvm.viewmodel to dev.ikm.komet.application, javafx.fxml, org.carlfx.cognitive;
 
-    provides dev.ikm.komet.framework.KometNodeFactory with dev.ikm.komet.kview.mvvm.view.details.DetailsNodeFactory, dev.ikm.komet.kview.mvvm.view.properties.PropertiesNodeFactory;
-
     exports dev.ikm.komet.kview.mvvm.view.confirmation;
     opens dev.ikm.komet.kview.mvvm.view.confirmation;
+    exports dev.ikm.komet.kview.mvvm.viewmodel.confirmation;
+    opens dev.ikm.komet.kview.mvvm.viewmodel.confirmation to dev.ikm.komet.application, javafx.fxml, org.carlfx.cognitive;
+
+    provides dev.ikm.komet.framework.KometNodeFactory with dev.ikm.komet.kview.mvvm.view.details.DetailsNodeFactory, dev.ikm.komet.kview.mvvm.view.properties.PropertiesNodeFactory;
 
     provides EntityKlWindowFactory with
             dev.ikm.komet.kview.klwindows.concept.ConceptKlWindowFactory,
